auth: Unconditionally create the `system_auth` keyspace
The `system_auth` keyspace is used to store tables for authentication
and authorization metadata.

Previously, this keyspace would only be created if either of the
non-default authenticator or authorizer were activated in configuration.

The upcoming role-management system is enabled unconditionally and also
uses the `system_auth` keyspace for its metadata.